Player stamina in HUD

+++ Assets/Brad/Scripts/DataModel/SaveData.cs
@@ -15,7 +15,7 @@ public class SaveData
     public float PLAYER_Health;
     public float PLAYER_MaxHealth;
     public int PLAYER_XP;
-    public int PLAYER_Stamina;
+    public float PLAYER_Stamina;
     public int PLAYER_CurrentAmmo;
 
     public string PLAYER_EquippedWeapon;
